refactor: dissolve api/properties.py — import Property and the registry directly
The module had shrunk to a re-export of Property plus
MANAGED_PROPERTY_KEYS, itself just frozenset(DELTA_PROPERTY_REGISTRY).
api/__init__ now re-exports Property straight from
application.properties, and DeltaTable's typo check reads the registry
directly (a registry lookup, same semantics). One fewer alias for the
same objects.
