+++ src/delta_engine/api/__init__.py
@@ -1,5 +1,5 @@
-from delta_engine.api.properties import Property
 from delta_engine.api.table import DeltaTable, ForeignKey, Self
+from delta_engine.application.properties import Property
 from delta_engine.domain.model import (
     Array,
     Boolean,
