+++ tests/api/test_table.py
@@ -1,8 +1,8 @@
 import pytest
 
 from delta_engine.api import Column, DeltaTable, ForeignKey, Integer, String
-from delta_engine.api.properties import Property
 from delta_engine.api.table import METADATA_ASPECTS
+from delta_engine.application.properties import Property
 from delta_engine.domain.model import (
     ALL_ASPECTS,
     Column as DomainColumn,
